Attribute tests and bug fixes, Collections tests and refactor, extensions test and fixes, Math tests + new normalize overload

+++ TPFramework.Core/Source/TPAttributePackage/TPModifierList.cs
@@ -18,7 +18,7 @@ namespace TPFramework.Core
         
         protected List<T> Modifiers { get; set; }
 
-        public float Count { get { return Modifiers.Count; } }
+        public int Count { get { return Modifiers.Count; } }
 
         public T this[int index] { get { return Modifiers[index]; } }
